// src/shared/enums/OdinIPC.ts
export enum OdinIPC {
  GET_SUPPLIERS = 'getSuppliers',
  GET_CATEGORIES = 'getCategories',
  SYNC_OPERATORS = 'syncOperators',
  GET_PRODUCT_BY_BARCODE = 'getProductByBarcode',
  GET_OPERATORS = 'getOperators',
  OPERATOR_LOGIN = 'operatorLogIn',
  OPERATOR_LOGOUT = 'operatorLogOut',
  SWITCH_OPERATOR = 'switchOperator',
  GET_PRODUCTS = 'getProducts',
  GET_PURCHASE_ORDERS = 'getPurchaseOrders',
  CLOSE_DAY = 'closeDay',
  GET_DAISY_EXPERT_PORT = 'getDaisyExpertPort',
  ORDER = 'order',
  PRE_ORDER = 'preOrder',
  CREATE_RETURN_ORDER = 'createReturnOrder',
  RETURN_ORDER = 'returnOrder',
  FINISH_RETURN_ORDER = 'finishReturnOrder',
  CANCEL_RETURN_ORDER = 'cancelReturnOrder',
  GET_WIFI_STATUS = 'getWifiStatus',
  GET_WIFI_NETWORKS = 'getWifiNetworks',
  CONNECT_TO_WIFI = 'connectToWifi',
  DISCONNECT_WIFI = 'disconnectWifi',
  GET_ENV = 'getEnv',
  IS_INITED = 'isInited',
  SET_ENV = 'setEnv',
  REGISTER_DEVICE = 'registerDevice',
  GET_DEVICE_ID = 'getDeviceId',
  CHECK_UPDATES = 'checkUpdates',
  REBOOT = 'reboot',
  INSTALL_UPDATE = 'installUpdate',
  SYNC = 'sync',
  GET_LAST_SYNC_DATE = 'getLastSyncDate',
  GET_APP_VER = 'getAppVer',
  RECEIVE_ORDER = 'receiveOrder',
  CHECK_ADMIN_CREDENTIALS = 'checkAdminCredentials',
  GET_OPERATOR = 'getOperator',
  SHUT_DOWN = 'shutDown',
  START_SHIFT = 'startShift',
  FINISH_SHIFT = 'finishShift',
  GET_SHIFT = 'getShift',
  SET_DEVICE_SN = 'setDeviceSerialNumber',
  SET_CONFIG = 'setConfig',
  ENTITY_CLIENTS = 'getEntityClients',
  ALL_ENTITY_CLIENTS = 'getAllEntityClients',
  ENTITY_ORDER = 'entityOrder',
  CHECK_PRIVILEGE_ELEVATION_PASSWORD = 'checkPrivilegeElevationPassword',
  REQUEST_PRIVILEGE_ELEVATION_BY_PASSWORD = 'requestPrivelegeElevationByPassword',
  SUBMIT_PRIVILEGE_ELEVATION_PASSWORD = 'submitPrivilegeElevationPassword',
  GET_IMEIS_BY_STOCK_ITEM = 'getIMEISByStockItem',
  WITHDRAW_CASH = 'withdrawCash',
  CHECK_SET_TABLE = 'check/setTable',
  CHECK_SET_OPERATOR = 'check/setOperator',
  CHECK_ORDER = 'check/order',
  CHECK_ENTITY_ORDER = 'check/entityOrder',
  CHECK_PREORDER = 'check/preOrder',
  CHECK_SET_ACTIVE = 'check/setActive',
  CHECK_CREATE = 'check/create',
  CHECK_GET_ACTIVE = 'check/getActive',
  CHECK_UNSET_ACTIVE = 'check/unsetActive',
  CHECK_ADD_ITEM = 'check/addItem',
  CHECK_UPDATE_ITEM = 'check/updateItem',
  CHECK_REMOVE_ITEM = 'check/removeItem',
  CHECK_REMOVE_ALL_ITEMS = 'check/removeAllItems',
  CHECK_DELETE = 'check/deleteCheck',
  CHECK_SET_ENTITY_DATA = 'check/setLegalEntityData',
  CHECK_UNSET_ENTITY_DATA = 'check/unsetLegalEntityData',
  CHECK_OPERATOR_HAS_CHECKS = 'check/operatorHasChecks',
  CHECK_UPDATE_GUEST_COUNT = 'check/updateGuestCount',
  CHECK_GET_CHECKS = 'check/getChecks',
  CHECK_HAS_ACTIVE_CHECKS = 'check/hasActiveChecks',
  CHECK_SET_TAX_RATE = 'check/setTaxRate',
  CLOCK_CHECK_DRIFT_STATUS = 'clock/checkStatus',
  CLOCK_WARNING_SHOWN = 'clock/warningShown',
  GET_SPACES = 'getSpaces',
  GET_TABLES_WITH_STATUS = 'getTablesWithStatus',
  GET_SHIFT_TOTAL_SALES = 'getShiftTotalSales',
  GET_TRANSACTION_HISTORY_LIST = 'getTransactionHistoryList',
  GET_TRANSACTION_DETAILS = 'getTransactionDetails',
}

export function OdinIPCResult(event: OdinIPC) {
  return `${event}.result`;
}