// src/shared/enums/OdinEvent.ts
export enum OdinEvent {
  SYNCED = 'event.synced',
  DAY_TO_BE_CLOSED = 'event.dayToBeClosed',
  RECEIVED_ORDER_SYNCED = 'event.receivedOrderSynced',
  UPDATE_DOWNLOADING = 'event.updateDownloading',
  UPDATE_DOWNLOADED = 'event.updateDownloaded',
  INTERNET_STATUS_CHANGED = 'event.internetStatusChanged',
  BOOT = 'event.boot',
  SUBSCRIBE = 'event.subscribe',
  CLOCK_DRIFT_STATUS = 'event.clockDriftStatus',
}